requirements.txt
asgiref==3.8.1
Django==5.0.6
django-bootstrap-form==3.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
prettyconf==2.2.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
setuptools==70.1.0
sqlparse==0.5.0
